# test: add the tableNodeComponent test