--- a/packages/ui/src/components/Node/TableNodeComponent.cy.tsx
+++ b/packages/ui/src/components/Node/TableNodeComponent.cy.tsx
@@ -5,22 +5,25 @@ import { ReactFlowProvider } from 'reactflow';
 import { normal } from './itemExamples/normal';
 import { nested } from './itemExamples/nested';
 import { oversize } from './itemExamples/oversize';
+import { createLargeRows } from './itemExamples/createLargeRows';
+import { createLargeColsFn } from './itemExamples/createLargeColsFn';
+
+const data = {
+  'params': [],
+  'computer': 'Table',
+  'label': 'Table',
+  'inputs': [
+    {
+      'id': 'Table.1.input',
+      'name': 'input',
+      'schema': {}
+    }
+  ],
+  'outputs': []
+};
+const id = 'Table.1';
 
 const mountTableNodeComponent = () => {
-  const data = {
-    'params': [],
-    'computer': 'Table',
-    'label': 'Table',
-    'inputs': [
-      {
-        'id': 'Table.1.input',
-        'name': 'input',
-        'schema': {}
-      }
-    ],
-    'outputs': []
-  };
-  const id = 'Table.1';
   cy.mount(
     <DataStoryProvider>
       <ReactFlowProvider>
@@ -30,97 +33,34 @@ const mountTableNodeComponent = () => {
   );
 }
 
-describe('test TableNodeComponent for table', () => {
-  function mockGetItems(items): void {
-    cy.stub(store, 'createStore').returns(() => {
-      return {
-        server: {
-          itemsApi: () => {
-            const getItems = ({
-              atNodeId = '',
-              limit = 10,
-              offset = 0
-            }) => {
-              return [items].slice(offset, offset + limit)
-            }
+let testPerformanceLimit = 20;
+const mockGetItems = (items: unknown[]): void => {
+  // Set the limit constant to 10000 for testing performance
 
-            return {
-              getItems
-            };
+  cy.stub(store, 'createStore').returns(() => {
+    return {
+      server: {
+        itemsApi: () => {
+          const getItems = ({
+            atNodeId = '',
+            limit = testPerformanceLimit,
+            offset = 0
+          }) => {
+            return items.slice(offset, offset + testPerformanceLimit)
           }
+
+          return {
+            getItems
+          };
         }
       }
-    });
-  }
-
-  it('render default component', () => {
-    mockGetItems([]);
-    mountTableNodeComponent();
-
-    cy.dataCy('data-story-table-node').should('exist');
+    }
   });
-
-  it('render component with data', () => {
-    mockGetItems(normal);
-    mountTableNodeComponent();
-
-    const headerLength = Object.keys(normal).length;
-
-    cy.dataCy('data-story-table-th').should('have.length',headerLength);
-    cy.dataCy('data-story-table-row').should('have.length', 1);
-  });
-
-  it('render component with nested data', () => {
-    mockGetItems(nested);
-    mountTableNodeComponent();
-
-    cy.dataCy('data-story-table-th').eq(0).should('have.text', 'objectId');
-    cy.dataCy('data-story-table-th').eq(1).should('have.text', 'properties.firstname');
-    cy.get('[data-cy="data-story-table-row"] > :nth-child(2)').should('have.text', 'John');
-  })
-
-  it('render component with oversize key or value data', () => {
-    mockGetItems(oversize);
-    mountTableNodeComponent();
-
-    cy.dataCy('data-story-table-th').eq(3).should('have.text', 'long_long_long_long_...');
-    cy.get('[data-cy="data-story-table-row"] > :nth-child(2)').should('have.text', 'long_long_long_long_...');
-  });
-
-  it('render component with line break data', () => {
-    mockGetItems(nested);
-    mountTableNodeComponent();
-
-    cy.dataCy('data-story-table-th').eq(7).should('have.text', 'address.street');
-    cy.get('[data-cy="data-story-table-row"] > :nth-child(8)').should('have.text', '122 Main St\nSuite 10...');
-  })
-})
+};
 
 describe('test TableNodeComponent for tooltip', () => {
-  function mockGetItems(items): void {
-    cy.stub(store, 'createStore').returns(() => {
-      return {
-        server: {
-          itemsApi: () => {
-            const getItems = ({
-              atNodeId = '',
-              limit = 10,
-              offset = 0
-            }) => {
-              return [items].slice(offset, offset + limit)
-            }
-
-            return {
-              getItems
-            };
-          }
-        }
-      }
-    });
-  }
-
   it('render tooltip with normal data', () => {
-    mockGetItems(normal);
+    mockGetItems([normal]);
     mountTableNodeComponent();
 
     cy.dataCy('data-story-table-th').eq(1).click();
@@ -128,7 +68,7 @@ describe('test TableNodeComponent for tooltip', () => {
   });
 
   it('render tooltip with oversize data', () => {
-    mockGetItems(oversize);
+    mockGetItems([oversize]);
     mountTableNodeComponent();
 
     // test long key on tooltip
@@ -143,7 +83,7 @@ describe('test TableNodeComponent for tooltip', () => {
   });
 
   it('render tooltip with line break data', () => {
-    mockGetItems(nested);
+    mockGetItems([nested]);
     mountTableNodeComponent();
 
     cy.get('[data-cy="data-story-table-row"] > :nth-child(8)').click();
@@ -152,3 +92,115 @@ Suite 100
  Anytown`);
   });
 });
+
+describe('test TableNodeComponent for table', () => {
+  it('render default component', () => {
+    mockGetItems([]);
+    mountTableNodeComponent();
+
+    cy.dataCy('data-story-table-node').should('exist');
+  });
+
+  it('render component with empty data', () => {
+    mockGetItems([]);
+    mountTableNodeComponent();
+
+    cy.dataCy('data-story-table-th').should('have.length', 1);
+    cy.dataCy('data-story-table-th').should('have.text', 'Awaiting data');
+  });
+
+  it('render component with data', () => {
+    mockGetItems([normal]);
+    mountTableNodeComponent();
+
+    const headerLength = Object.keys(normal).length;
+
+    cy.dataCy('data-story-table-th').should('have.length',headerLength);
+    cy.dataCy('data-story-table-row').should('have.length', 1);
+  });
+
+  it('render component with nested data', () => {
+    mockGetItems([nested]);
+    mountTableNodeComponent();
+
+    cy.dataCy('data-story-table-th').eq(0).should('have.text', 'objectId');
+    cy.dataCy('data-story-table-th').eq(1).should('have.text', 'properties.firstname');
+    cy.get('[data-cy="data-story-table-row"] > :nth-child(2)').should('have.text', 'John');
+  })
+
+  it('render component with oversize key or value data', () => {
+    mockGetItems([oversize]);
+    mountTableNodeComponent();
+
+    cy.dataCy('data-story-table-th').eq(3).should('have.text', 'long_long_long_long_...');
+    cy.get('[data-cy="data-story-table-row"] > :nth-child(2)').should('have.text', 'long_long_long_long_...');
+  });
+
+  it('render component with line break data', () => {
+    mockGetItems([nested]);
+    mountTableNodeComponent();
+
+    cy.dataCy('data-story-table-th').eq(7).should('have.text', 'address.street');
+    cy.get('[data-cy="data-story-table-row"] > :nth-child(8)').should('have.text', '122 Main St\nSuite 10...');
+  });
+
+  it('render component with 1000 columns data', () => {
+    const thousandCols = createLargeColsFn(1000);
+    mockGetItems([thousandCols]);
+    const start = performance.now();
+    mountTableNodeComponent();
+
+    // https://web.dev/articles/rail
+    cy.dataCy('data-story-table-th')
+      .eq(1)
+      .should('exist')
+      .then(() => {
+        const end = performance.now();
+        const renderTime = end - start;
+        cy.log(`cypress render Time: ${renderTime}ms`);
+      });
+
+    const headerLength = Object.keys(thousandCols).length;
+    cy.dataCy('data-story-table-th')
+      .should('have.length', headerLength);
+  });
+
+  it('render component with 1000 rows data', () => {
+    const thousandRows = createLargeRows(1000);
+    testPerformanceLimit = thousandRows.length;
+    mockGetItems(thousandRows);
+    const start = performance.now();
+    mountTableNodeComponent();
+
+    cy.dataCy('data-story-table-row')
+      .eq(1)
+      .should('exist')
+      .then(() => {
+        const end = performance.now();
+        const renderTime = end - start;
+        cy.log(`cypress render Time: ${renderTime}ms`);
+      });
+
+    cy.dataCy('data-story-table-row').should('have.length', testPerformanceLimit);
+  });
+
+  it('test table component scroll', () => {
+    const thousandRows = createLargeRows(1000);
+
+    testPerformanceLimit = thousandRows.length;
+    mockGetItems(thousandRows);
+    mountTableNodeComponent();
+
+    // 等待 10ms，等待数据加载完成
+    cy.wait(10);
+
+    const start = performance.now();
+    cy.dataCy('data-story-table')
+      .scrollTo('bottom')
+      .then(() => {
+        const end = performance.now();
+        const scrollTime = end - start;
+        cy.log(`cypress scroll Time: ${scrollTime}ms`);
+      });
+  });
+})