- **add the tableNodeComponent test**

The time it takes to render `N` rows of data at once and scroll to the bottom

| data (rows) |scroll to bottom time (ms) |
| ----------- | ----------------------- |
| 1000        | 2,010                  |
| 10,000       | 18,857                |
| 20,000       | 32,289                 |
| 30,000       | can't render             |

FPS data when manually scrolling to render `N` rows of data with `limit` set to 20.

| data (rows) | fps performance during table scrolling (fps)  |
| ----------- | ------- |
| 1000        | 54   |
| 10,000       | 42.6 |
| 100,000      | 25.8 |

![image](https://github.com/ajthinking/data-story/assets/20497176/f0629632-c762-41f3-9388-507cd9d03435)

- **feat: made the tooltip selectable and click**
before:

https://github.com/ajthinking/data-story/assets/20497176/80441828-9523-44ea-9c22-9ed1d44301e9

after:

https://github.com/ajthinking/data-story/assets/20497176/edad2937-9df4-478f-9eee-58ef36649069

- **enhanced the unit tests for ItemCollection**