@stone-lyl 2024-04-01
> Thank you 👍
> 
> Nice stress testing! 1000 items, like in the tests, is good enough for now.
> 
> I realized I was breaking my own rule here - please inline any itemExamples, smaller helper functions etc. And drop itemExamples dir.

I'll integrate the contents of the itemExamples directory into the mock.ts file to ensure our project structure remains consistent.

![image](https://github.com/ajthinking/data-story/assets/20497176/b5399402-d4a1-487e-a0ea-a3233e9fba62)